pyproject.toml
[tool.poetry]
name = "bot-builder-nlp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "1.60.0"
grpcio-tools = "1.60.0"
pymilvus = "^2.3.6"
pydantic-settings = "^2.2.0"
sentence-transformers = "^2.3.1"
redis = "^5.0.2"
levenshtein = "^0.25.0"
black = "^24.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "bot_builder_nlp.main:main"